Refactor, refactor, refactor!
Refactored the NodeInterface to be a base class, built NodeInterfaceString and NodeInterfaceSocket on top of the base class.

Added in some testing of the ChannelController via the NodeInterfaceString (built for use with testing)

Updated some of the parameter names, because I coded them wrong.

Added in the JSON testing library.